+++ .eslintrc.js
@@ -5,11 +5,22 @@ module.exports = {
     "browser": true,
     "jest": true
   },
-  "extends": "airbnb",
+  "extends": [
+    "airbnb",
+  ],
   "globals": {
     "__DEV__": true
   },
+  "plugins": [
+    "@typescript-eslint"
+  ],
   "settings": {
+    "import/extensions": [
+      ".js",
+      ".jsx",
+      ".ts",
+      ".tsx"
+    ],
     "import/resolver": {
       "node": {
         "extensions": [
@@ -22,20 +33,50 @@ module.exports = {
     }
   },
   "rules": {
-    "max-len": [
-      1,
-      140,
-      2
-    ],
     camelcase: [
       "error",
       {
         allow: [
           "^oauth_",
-          "include_entities",
+          "^include_",
+          "^default_profile",
+          "^profile_",
+          "^withheld_",
+          "_count$",
+          "_url$",
           "skip_status",
-          "include_email"
+          "screen_name",
+          "user_id",
+          "created_at",
+          "id_str"
         ]
+      }
+    ],
+    "no-unused-vars": "off",
+    "@typescript-eslint/no-unused-vars": "error",
+    'react/jsx-props-no-spreading': ['error', {
+      html: 'enforce',
+      custom: 'ignore',
+      exceptions: [],
+    }],
+    "max-len": [
+      1,
+      140,
+      2
+    ],
+    "jsx-a11y/label-has-for": [2, {
+      "required": {
+        "some": ["nesting", "id"]
+      }
+    }],
+    "react/prop-types": [
+      0
+    ],
+    "react/destructuring-assignment": 0,
+    "react/jsx-one-expression-per-line": [
+      0,
+      {
+        "allow": "literal"
       }
     ],
     "import/no-extraneous-dependencies": [
@@ -46,7 +87,6 @@ module.exports = {
         "peerDependencies": false
       }
     ],
-    "react/jsx-props-no-spreading": 0,
     "react/jsx-filename-extension": [
       "error",
       {
@@ -58,6 +98,14 @@ module.exports = {
         ]
       }
     ],
-    "react/destructuring-assignment": [0]
+    "import/extensions": [
+      "error", "always",
+      {
+        "js": "never",
+        "jsx": "never",
+        "ts": "never",
+        "tsx": "never"
+      }
+    ]
   }
 }