+++ src/client.ts
@@ -1,6 +1,13 @@
 /* lib */
 import Request from './request';
 import * as Util from './util';
+import {
+  Method,
+  RequestTokenResponse,
+  AccessTokenResponse,
+  AccessToken,
+} from './types';
+import CustomError from './error';
 
 /* const */
 const baseURL: string = 'https://api.twitter.com';
@@ -43,11 +50,16 @@ class Client {
     this.TokenRequestHeaderParams = Util.createTokenRequestHeaderParams(this.ConsumerKey, { callback });
     this.TokenRequestHeaderParams = Util.createSignature(this.TokenRequestHeaderParams, 'POST', baseURL + requestTokenURL, this.ConsumerSecret);
 
-    const result = await Request(
+    const result = await Request<RequestTokenResponse>(
       'POST',
       baseURL + requestTokenURL,
       this.TokenRequestHeaderParams,
     );
+
+    if ('errors' in result) {
+      throw new CustomError(result);
+    }
+
     this.setAccessToken(result.oauth_token, result.oauth_token_secret);
 
     return `${baseURL + authorizationURL}?oauth_token=${this.Token}`;
@@ -56,16 +68,21 @@ class Client {
   /**
    * get access token
    */
-  getAccessToken = async (verifier: string = ''): Promise<{ errors?: any, oauth_token: string, oauth_token_secret: string }> => {
+  getAccessToken = async (verifier: string = ''): Promise<AccessToken> => {
     this.TokenRequestHeaderParams = Util.createTokenRequestHeaderParams(this.ConsumerKey, { token: this.Token });
     this.TokenRequestHeaderParams = Util.createSignature(this.TokenRequestHeaderParams, 'POST', baseURL + accessTokenURL, this.ConsumerSecret, this.TokenSecret);
     this.TokenRequestHeaderParams.oauth_verifier = verifier;
 
-    const result = await Request(
+    const result = await Request<AccessTokenResponse>(
       'POST',
       baseURL + accessTokenURL,
       this.TokenRequestHeaderParams,
     );
+
+    if ('errors' in result) {
+      throw new CustomError(result);
+    }
+
     this.setAccessToken(result.oauth_token, result.oauth_token_secret);
 
     return { oauth_token: result.oauth_token, oauth_token_secret: result.oauth_token_secret };
@@ -74,24 +91,28 @@ class Client {
   /**
    * call Twitter Api
    */
-  api = async (method: string = 'GET', endpoint: string, params: any = {}): Promise<any> => {
-    const apiMethod = method.toUpperCase();
+  api = async <T>(method: Method, endpoint: string, params: any = {}): Promise<T> => {
     const apiEndpoint = endpoint.slice(0, 1) !== '/' ? `/${endpoint}` : endpoint;
 
     this.TokenRequestHeaderParams = Util.createTokenRequestHeaderParams(this.ConsumerKey, { token: this.Token, params });
-    this.TokenRequestHeaderParams = Util.createSignature(this.TokenRequestHeaderParams, apiMethod, apiURL + apiEndpoint, this.ConsumerSecret, this.TokenSecret);
+    this.TokenRequestHeaderParams = Util.createSignature(this.TokenRequestHeaderParams, method, apiURL + apiEndpoint, this.ConsumerSecret, this.TokenSecret);
 
-    const result = await Request(
-      apiMethod,
+    const result = await Request<T>(
+      method,
       apiURL + (params ? `${apiEndpoint}?${Util.encodeParamsToString(params)}` : apiEndpoint),
       this.TokenRequestHeaderParams,
     );
+
+    if ('errors' in result) {
+      throw new CustomError(result);
+    }
 
     return result;
   }
 
   /**
    * api("POST",endpoint,params) alias
+   * will be remove at next version
    */
   post = async (endpoint: string, params: any = {}): Promise<any> => {
     const result = await this.api('POST', endpoint, params);
@@ -101,6 +122,7 @@ class Client {
 
   /**
    * api("GET",endpoint,params) alias
+   * will be remove at next version
    */
   get = async (endpoint: string, params: any = {}): Promise<any> => {
     const result = await this.api('GET', endpoint, params);