+++ CHANGELOG.md
@@ -1,5 +1,23 @@
 # Change Log
 
+## v3.0.0 - 2019-12-18
+### Added
+- custom hook add `useTwitter`
+- api `Method` type
+- try catch error
+- custom error
+
+### Changed
+- types
+
+### Removed
+- api's method to UpperCase()
+
+### Will be removed next version
+- `twitter.get()`
+- `twitter.post()`
+- `TWLoginButton`
+
 ## v2.4.0 - 2019-10-16
 ### Added
 - TWLoginButton add `props.closeText`